Move UIImageView.optimizedURL to SiteIconViewModel
